apps/ml-service/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
psycopg2-binary
gunicorn
spacy
numpy
pandas
